+++ lib/react_on_rails_pro/server_rendering_pool/node_rendering_pool.rb
@@ -41,18 +41,20 @@ module ReactOnRailsPro
             .exec_server_render_js(js_code, render_options, self)
         end
 
+        def exec_server_render_streaming_js(js_code, render_options)
+          render_options.set_option(:throw_js_errors, ReactOnRailsPro.configuration.throw_js_errors)
+          # The secret sauce is passing self as the 3rd param, the js_evaluator
+          ReactOnRails::ServerRenderingPool::RubyEmbeddedJavaScript
+            .exec_server_render_streaming_js(js_code, render_options, self)
+        end
+
+        def eval_streaming_js(js_code, render_options)
+          path = prepare_render_path(js_code, render_options)
+          ReactOnRailsPro::Request.render_code_as_stream(path, js_code)
+        end
+
         def eval_js(js_code, render_options, send_bundle: false)
-          ReactOnRailsPro::ServerRenderingPool::ProRendering
-            .set_request_digest_on_render_options(js_code, render_options)
-
-          # In case this method is called with simple, raw JS, not depending on the bundle, next line
-          # is needed.
-          @bundle_hash ||= ReactOnRailsPro::Utils.bundle_hash
-
-          # TODO: Remove the request_digest. See https://github.com/shakacode/react_on_rails_pro/issues/119
-          # From the request path
-          # path = "/bundles/#{@bundle_hash}/render"
-          path = "/bundles/#{@bundle_hash}/render/#{render_options.request_digest}"
+          path = prepare_render_path(js_code, render_options)
 
           response = ReactOnRailsPro::Request.render_code(path, js_code, send_bundle)
 
@@ -66,12 +68,26 @@ module ReactOnRailsPro
             raise ReactOnRailsPro::Error,
                   "Renderer unhandled error at the VM level: #{response.code}:\n#{response.body}"
           else
-            raise ReactOnRailsPro::Error, "Unknown response code from renderer: #{response.code}:\n#{response.body}"
+            raise ReactOnRailsPro::Error, "Unexpected response code from renderer: #{response.code}:\n#{response.body}"
           end
         rescue StandardError => e
           raise e unless ReactOnRailsPro.configuration.renderer_use_fallback_exec_js
 
           fallback_exec_js(js_code, render_options, e)
+        end
+
+        def prepare_render_path(js_code, render_options)
+          ReactOnRailsPro::ServerRenderingPool::ProRendering
+            .set_request_digest_on_render_options(js_code, render_options)
+
+          # In case this method is called with simple, raw JS, not depending on the bundle, next line
+          # is needed.
+          @bundle_hash ||= ReactOnRailsPro::Utils.bundle_hash
+
+          # TODO: Remove the request_digest. See https://github.com/shakacode/react_on_rails_pro/issues/119
+          # From the request path
+          # path = "/bundles/#{@bundle_hash}/render"
+          "/bundles/#{@bundle_hash}/render/#{render_options.request_digest}"
         end
 
         def fallback_exec_js(js_code, render_options, error)