+++ spec/dummy/app/controllers/pages_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class PagesController < ApplicationController
+  include ActionController::Live
+
   XSS_PAYLOAD = { "<script>window.alert('xss1');</script>" => '<script>window.alert("xss2");</script>' }.freeze
   PROPS_NAME = "Mr. Server Side Rendering"
   APP_PROPS_SERVER_RENDER = {
@@ -10,6 +12,7 @@ class PagesController < ApplicationController
   }.freeze
 
   include ReactOnRails::Controller
+  include ReactOnRailsPro::Stream
 
   before_action do
     session[:something_useful] = "REALLY USEFUL"
@@ -20,8 +23,24 @@ class PagesController < ApplicationController
   before_action :initialize_shared_store, only: %i[client_side_hello_world_shared_store_controller
                                                    server_side_hello_world_shared_store_controller]
 
+  # Used for testing streamed html pages
+  # Capybara doesn't support streaming, so we need to navigate to an empty page first
+  # and then make an XHR request to the desired page
+  # We need to navigate to an empty page first to avoid CORS issues and to update the page host
+  def empty
+    render plain: ""
+  end
+
   def cached_react_helmet
     render "/pages/pro/cached_react_helmet"
+  end
+
+  def stream_async_components
+    stream_view_containing_react_components(template: "/pages/stream_async_components")
+  end
+
+  def stream_async_components_for_testing
+    stream_view_containing_react_components(template: "/pages/stream_async_components_for_testing")
   end
 
   def loadable_component